+++ backend/api/users/tests/test_request_reset_password.py
@@ -1,4 +1,4 @@
-from emails.templates import EmailTemplate
+from notifications.templates import EmailTemplate
 import pytest
 
 from users.tests.factories import UserFactory
