+++ backend/grants/tasks.py
@@ -3,7 +3,7 @@ from urllib.parse import urljoin
 
 from django.conf import settings
 from django.utils import timezone
-from emails.templates import EmailTemplate
+from notifications.templates import EmailTemplate
 
 from users.models import User
 from grants.models import Grant
