+++ backend/notifications/templates.py
