+++ backend/schedule/tasks.py
@@ -7,12 +7,11 @@ from google_api.exceptions import NoGoogleCloudQuotaLeftError
 from googleapiclient.errors import HttpError
 from google_api.sdk import youtube_videos_insert, youtube_videos_set_thumbnail
 from integrations import plain
-from notifications.emails import mark_safe
 from pretix import user_has_admission_ticket
 from django.utils import timezone
 from grants.tasks import get_name
-from emails.templates import EmailTemplate
-from notifications.emails import send_email
+from notifications.templates import EmailTemplate
+from notifications.emails import send_email, mark_safe
 from urllib.parse import urljoin
 from django.conf import settings
 import logging
